Moves upgrade-series test concerns into a new file.
